Fix: badge: uninitialized message direction used to enqueue messages
The right-most badge sometimes fails to pair with the following error
log: "Failed to enqueue message: attempted to send a message out of bounds".

This occurs when, for example, a badge at the right-most end of a chain
attempts to send a message to its right-hand neighbor. This is
necessarily an internal error.

_direction is not initialized for left-most and right-most badges since
the direction to use is constant: the left-most badge should always send
its messages to the right, and the right-most badge should always send
its messages to the left.

The use of _direction is replaced by peer_relative_position::LEFT.

The direction field is changed to no-longer be a bitfield as it doesn't
really matter for this platform. However, Michael and Sylvain had both
reported that changing the fields before and after _direction from
bitfields to full-size fields seemed to affect the pairing success rate.

Since _direction is not initialized, there's a good chance the flakyness
was caused/affected by the layout of the class' members. Moreover,
initializing _direction to 0 would give the correct behaviour of sending
messages to the left which probably hid the problem.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@gmail.com>