Fix: badge: message sent confirmation received before state transition
The id_exchanger relies on the message_sent event to send the badge's ID
from the left-most badge. As such, the id_exchanger must have the focus
before the ANIMATION_COMPLETE message is enqueued to ensure it receives
and handles the message sent confirmation.

Otherwise, the pairing never completes as the badges wait for the
left-most peer to send its unique ID annoucement message.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@gmail.com>