Badge - cleanup and update (#576)
* strip_animator: remove unused code

Removed unused code from the 2023/2024 strip animator badge code.

Signed-off-by: Sylvain Lemieux <sylemieux.git@gmail.com>

* strip_animator: add specific logger config

Add support to be able to control the logging capability of the
strip animator individually.

Signed-off-by: Sylvain Lemieux <sylemieux.git@gmail.com>

* config_store: fix spacing alignment

Signed-off-by: Sylvain Lemieux <sylemieux.git@gmail.com>

* config: define "social level check" field name

Add "social level check" field name to use in the persistence module.

Signed-off-by: Sylvain Lemieux <sylemieux.git@gmail.com>

* config_store: add "social level check" to persistence storage

Add the "social level check" value to "config" persistence storage.

Signed-off-by: Sylvain Lemieux <sylemieux.git@gmail.com>

* badge: minimum  change to work with latest "config store"

Add minimum changes to work with the latest "config store"
that include a 32 bits value for "social level check".

Signed-off-by: Sylvain Lemieux <sylemieux.git@gmail.com>

* Add Copyright - 2025 NorthSec

Signed-off-by: Sylvain Lemieux <sylemieux.git@gmail.com>

* badge: update config load logging

Log the config load detected information on multiple lines
instead of only one.

Signed-off-by: Sylvain Lemieux <sylemieux.git@gmail.com>

* badge: add "social level check" support

For the 2023 & 2024 badges, only the "social level" was written to NVS memory.
This implementation was hacked during the 2023 conference.

Add a very basic 32 bits obfuscation derived from the MAC address & the social level.
This is to ensure another value is written to the NVS memory with the social level.
If the social level check failed, the social is reset to 0.

This change does not add any value to the software after the conference
(when the source code is release).

Signed-off-by: Sylvain Lemieux <sylemieux.git@gmail.com>

* badge: add boundary check when new sponsor apply

Signed-off-by: Sylvain Lemieux <sylemieux.git@gmail.com>

---------

Signed-off-by: Sylvain Lemieux <sylemieux.git@gmail.com>
Co-authored-by: Sylvain Lemieux <sylemieux.git@gmail.com>