network handler: adding delay to allow "complete" display on LCD
With the previous commit that display the "New badge" / "No new badge"
on LCD, when sync is completed, the "Complete" status is no longer
visible on the badge, when the idle screen is #1.

Adding a delay to ensure the "Complete" status is always visible
before sending the result of the sync to the badge module.

Signed-off-by: Sylvain Lemieux <sylemieux.git@gmail.com>